==> 1_import_data.Rmd <==
---
title: "Import data into R"
output: html_document
---

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

There are many ways to import data into R, and numerous packages are available to analyze metabarcoding data. Among these, [Phyloseq](http://www.r-project.org) remains a versatile tool to import, store and analyze data. Moreover, many other packages can use a Phyloseq object as a starting point. We'll start by importing the tables using the Phyloseq package. 

For these steps wee will need the following libraries: 

```{r, message=FALSE,echo=TRUE}
library("tidyverse")
library("phyloseq")
```

Tables to Phyloseq
----
Let's start by importing the tables from the raw_data directory: 
- using the read.csv function import the table;
- check the first rows and columns with the head() function;
- define the row names from the OTU column with tibble::column-to_rownames;
- transform otu and taxonomy tables into matrixes (and confirm the object class with the class() function)

```{r}
biom_illumina<- read.csv("Raw_data/Illumina/illumina_ASV_table.tsv", 
                        header=TRUE, sep="\t")
head(biom_illumina[,1:5],10)

biom_illumina <- biom_illumina %>%
  tibble::column_to_rownames("otu") 

biom_illumina <- as.matrix(biom_illumina)
class(biom_illumina)
```

Next, import the taxonomy table in the same way.

```{r}
taxo_illumina<- read.csv("Raw_data/Illumina/illumina_taxonomy.tsv", 
                        header=TRUE, sep="\t")
head(taxo_illumina[,1:8],10)

taxo_illumina <- taxo_illumina %>%
  tibble::column_to_rownames("otu")

taxo_illumina <- as.matrix(taxo_illumina)
class(taxo_illumina) 
```

And merge the two tables into a single phyloseq objects

```{r}
OTU_illumina = phyloseq::otu_table(biom_illumina, taxa_are_rows = TRUE)
TAX_illumina = phyloseq::tax_table(taxo_illumina)

illuminaPS <- phyloseq(OTU_illumina, TAX_illumina)
illuminaPS
```

Let's repeat for the IonTorrent dataset 

```{r}
biom_IonTorrent<- read.csv("Raw_data/IonTorrent/iontorrent_ASV_table.tsv", 
                        header=TRUE, sep="\t")
head(biom_IonTorrent[,1:5],10)

biom_IonTorrent <- biom_IonTorrent %>%
  tibble::column_to_rownames("otu")

biom_IonTorrent <- as.matrix(biom_IonTorrent)
class(biom_IonTorrent)


taxo_IonTorrent<- read.csv("Raw_data/IonTorrent/iontorrent_taxonomy.tsv", 
                        header=TRUE, sep="\t")
head(taxo_IonTorrent[,1:8],10)

taxo_IonTorrent <- taxo_IonTorrent %>%
  tibble::column_to_rownames("otu") 

taxo_IonTorrent <- as.matrix(taxo_IonTorrent)
class(taxo_IonTorrent) 

OTU_IonTorrent = phyloseq::otu_table(biom_IonTorrent, taxa_are_rows = TRUE)
TAX_IonTorrent = phyloseq::tax_table(taxo_IonTorrent)

IonTorrentPS <- phyloseq(OTU_IonTorrent, TAX_IonTorrent)
IonTorrentPS
```

And Again for the Nanopore dataset 

```{r}
biom_Nanopore<- read.csv("Raw_data/Nanopore/Nanopore_pseudo_ASV_table.csv", 
                        header=TRUE, sep=";")
head(biom_Nanopore[,1:5],10)

biom_Nanopore <- biom_Nanopore %>%
  tibble::column_to_rownames("otu")

biom_Nanopore <- as.matrix(biom_Nanopore)
class(biom_Nanopore)


taxo_Nanopore<- read.csv("Raw_data/Nanopore/Nanopore_taxonomy.csv", 
                        header=TRUE, sep=";")
head(taxo_Nanopore[,1:8],10)

taxo_Nanopore <- taxo_Nanopore %>%
  tibble::column_to_rownames("otu") 

taxo_Nanopore <- as.matrix(taxo_Nanopore)
class(taxo_Nanopore) 

OTU_Nanopore = phyloseq::otu_table(biom_Nanopore, taxa_are_rows = TRUE)
TAX_Nanopore = phyloseq::tax_table(taxo_Nanopore)

NanoporePS <- phyloseq(OTU_Nanopore, TAX_Nanopore)
NanoporePS
```


Merging objects
----

Now we want to merge the phyloseq objects we've created
To do that we will generate:
- a combined biom-table (using the phyloseq::otu_table function)
- a combined taxonomy table (using the phyloseq::tax_table)  

```{r}
otu_illumina = phyloseq::otu_table(illuminaPS)
otu_Iontorrent = phyloseq::otu_table(IonTorrentPS)
otu_Nanopore = phyloseq::otu_table(NanoporePS)

combined_otu <- merge_phyloseq(otu_illumina, otu_Iontorrent, otu_Nanopore)
```


```{r}
tax_illumina  = phyloseq::tax_table(illuminaPS)
tax_Iontorrent = phyloseq::tax_table(IonTorrentPS)
tax_Nanopore = phyloseq::tax_table(NanoporePS)

combined_tax <- merge_phyloseq(tax_illumina, tax_Iontorrent, tax_Nanopore)
```

Adding metadata 
----



```{r}
meta <- read.csv("Raw_data/metadata_organicsoil.tsv", 
                 header=TRUE, sep="\t")
head(meta)

meta <- meta %>% 
  tibble::column_to_rownames("SampleID")

samples = sample_data(meta)
```

Finally, merge the biom table, tax table and metadata into a final phyloseq object: 

```{r}
combinedPS <- merge_phyloseq(combined_otu, combined_tax, samples);combinedPS

```


```{r}
library("ape")

random_tree_combinedPS = rtree(ntaxa(combinedPS), 
                    rooted=TRUE, 
                    tip.label=taxa_names(combinedPS))


combinedPS_final <- phyloseq(combined_otu, combined_tax, samples, random_tree_combinedPS)
combinedPS_final

```
Saving phyloseq objects
----

Before saving our work, to make sure that everything is under control, you can inspect the phyloseq object with the following functions: 

```{r message=FALSE,echo=TRUE, results="hide"}
sample_names(combinedPS_final)
rank_names(combinedPS_final)
sample_variables(combinedPS_final)
```

Finally, we save our phyloseq object into the processed_data directory
```{r}
saveRDS(combinedPS_final, file= "Processed_data/combinedPS_final.rds")
```
