==> 7_Time_series.Rmd <==
---
title: "Temporal diversity"
output: html_document
---

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```


```{r, message=FALSE,echo=TRUE}
library("ampvis2")
library("ggpubr")

WorkingPS_glom<-readRDS("Processed_data/WorkingPS_glom.rds")
```

Heat map by week and technologies  
---- 

First we need to convert our phyloseq object into an Ampvis2 object, then generate the heat map with the weeks (times) in x and the sequencing technology as facets. 

```{r, fig.align="center", echo = TRUE, fig.width = 14, fig.height= 6}
# Convert the PS object to ampvis 
WorkingPS_amp <- amp_load(WorkingPS_glom)
```


Time series
---- 

When data are collected over time, such as our aerobiome data, we may be tempted to explore the temporal distribution patterns of specific species. 
Continuing with the Ampvis2 package, we can use the amp_timeseries function to plot read distribution over time. 

```{r, message=FALSE,echo=TRUE}

amp_timeseries(WorkingPS_amp,
  time_variable = "Date",
  tax_aggregate = "Species",
  tax_show = 9,
  normalise=FALSE
)
```

Or with read counts to be in percent per sample.

```{r, message=FALSE,echo=TRUE}

amp_timeseries(WorkingPS_amp,
  time_variable = "Date",
  tax_aggregate = "Species",
  tax_show = 9,
  normalise=TRUE
)
```

We can also generate our curves by species in individual graphs. 

```{r, fig.align="center", echo = TRUE, fig.width = 14, fig.height= 6}

amp_timeseries(WorkingPS_amp,
  time_variable = "Date",
  #group_by = "seq_tech",
  split = TRUE,
  scales = "free_y",
  tax_show = 12,
  tax_aggregate = "Species",
  tax_add = NULL,
  normalise=TRUE,
  #plotly = TRUE,
  tax_empty="OTU"
)
```

Time series with error bars
---- 

Now let's try reproducing this figure manually to add an error bar (Standard error). We'll start again with the code we used for the stacked bar charts. 
First, let's rework the list of predefined species to match that of the graph obtained with Ampvis2, and replace anything not in our list with "Other":

```{r}

get_cols <- function (n){
  col <- c("#8dd3c7", "#ffffb3", "#bebada", "#fb8072", "#80b1d3",
           "#fdb462", "#b3de69", "#fccde5", "#d9d9d9", "#bc80bd",
           "#ccebc5", "#ffed6f")
  
  col2 <- c("#1f78e4", "#ffff33", "#c2a5cf", "#ff7f00", "#810f7c",
            "#a6cee3", "#006d2c", "#4d4d4d", "#8c510a", "#d73027",
            "#78c679", "#7f0000", "#41b6c4", "#e7298a", "#54278f")
  
  col3 <- c("#a6cee3", "#1f78b4", "#b2df8a", "#33a02c", "#fb9a99",
            "#e31a1c", "#fdbf6f", "#ff7f00", "#cab2d6", "#6a3d9a",
            "#ffff99", "#b15928")
  colorRampPalette(col2)(n)
}


library(tidyverse)
WorkingPS_melt <- psmelt(WorkingPS_glom)

head(WorkingPS_melt)

predefined_species = c("Peronospora manshurica",
                       "Peronospora variabilis",
                       "Bremia lactucae cluster",
                       "Hyaloperonospora nesliae",
                       "Peronospora destructor",
                       "Hyaloperonospora camelinae",
                       "Albugo candida",
                       "Peronospora aparines",
                       "Peronospora erucastri",
                       "Plasmopara viticola",
                       "Peronospora boni-henrici cluster"
)

WorkingPS_melt2 <-  WorkingPS_melt %>%
  mutate(Species = case_when(
    Species %in% 
      predefined_species ~ Species,  # Keep species from the list
    TRUE ~ "Other"  # Replace other species with "other"
  ))

```

Now we need to add a column for standard error (SE) in addition to relative abundance (rel_abund).

```{r}
genus_abundance2 <- WorkingPS_melt2 %>%
  group_by(Week2) %>%
  mutate(rel_abund = Abundance / sum(Abundance)*100) %>%
  mutate(SE = sd(rel_abund)/sqrt(length(rel_abund))) %>%
  ungroup() %>%
  select(-Abundance) %>%
  pivot_longer(c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
               names_to="level",
               values_to="taxon")%>%
  mutate(taxon = factor(taxon,
                       levels=c("Peronospora manshurica",
                       "Peronospora variabilis",
                       "Other",
                       "Bremia lactucae cluster",
                       "Hyaloperonospora nesliae",
                       "Peronospora destructor",
                       "Hyaloperonospora camelinae",
                       "Albugo candida",
                       "Peronospora aparines",
                       "Peronospora erucastri",
                       "Plasmopara viticola",
                       "Peronospora boni-henrici cluster"
                       )))
```

Then, summarize the data at the species level and pipe that into ggplot2 to get our line plot for the predefined species with error bars. 

```{r,  fig.align="center", echo = TRUE, fig.width = 14, fig.height= 6}
genus_abundance2 %>%
  filter(level=="Species") %>% 
  group_by(Week2, taxon) %>% 
  summarize(rel_abund = sum(rel_abund), SE=sum(SE)) %>%
  group_by(Week2, taxon) %>% 
  summarize(mean_rel_abund=mean(rel_abund), SE = mean(SE), .groups="drop") %>%
  ggplot(aes(x=Week2, y=mean_rel_abund, fill=taxon))+
  geom_point()+
  geom_errorbar(aes(ymin=mean_rel_abund-SE, ymax=mean_rel_abund+SE), width=0.2)+
  geom_line()+
  facet_wrap(vars(taxon), nrow = 3, scales="free")+
  labs(x="Week",
       y="Relative abundance (%)")+
  theme_bw()+
  theme(legend.position="none")
```

Alpha diversity over time 
---- 

In addition to the distribution of the relative abundance of one or more species over time, we may also be interested in looking at the evolution of alpha diversity as a function of time.

```{r,  fig.align="center", echo = TRUE, fig.width = 14, fig.height= 4}
library("MicrobiotaProcess")
alpha_oom <- get_alphaindex(WorkingPS_glom)

head(as.data.frame(alpha_oom))

Observe_status <- ggbox(alpha_oom, 
                        geom="boxplot", 
                        factorNames="Week",   
                        compare = FALSE,
                        testmethod = "wilcox.test",
                        signifmap = FALSE,
                        indexNames="Observe")+ 
  theme( legend.position="none")+
  theme(strip.background = element_rect(colour=NA, fill="grey"))+
  scale_fill_manual(values=get_cols(16))


Shannon_status<- ggbox(alpha_oom, 
                       geom="boxplot", 
                       factorNames="Week",   
                       compare = FALSE, 
                       testmethod = "wilcox.test",
                       signifmap = FALSE,
                       indexNames="Shannon")+ 
  theme(legend.position="Bottom")+
  theme(strip.background = element_rect(colour=NA, fill="grey"))+
  scale_fill_manual(values=get_cols(16))


library("ggpubr")
ggarrange(ggarrange(Observe_status,
                    Shannon_status,
                    ncol = 2, labels = c("A", "B"),
                    heights = c(2.5, 1)))

```
