==> 8_biomarker discovery.Rmd <==
---
title: "Indicator species"
author: "Hervé Van der Heyden"
date: "2024-10-25"
output: html_document
---

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```


Linear discriminant effect size analysis (LEfSe)
----

LEfSe analysis can be use to determine which species might explain the differences between groups. 
This approach uses a non-parametric Kruskal–Wallis test to detect species with significant differential abundance relative to the groups. 
Then, biological significance is tested using pairwise Wilcoxon's rank-sum tests, with the Benjamini–Hochberg correction to control the false discovery rate (FDR). 
In the final step, a linear discriminant analysis (LDA) is used to estimate the effect size of each differentially abundant feature.


```{r}

library("coin")
library("stringr")
library("ggtree") # BiocManager::install("ggtree")
library("ggtreeExtra") #install.packages("ggExtra")
library("MicrobiotaProcess")
library("tidyverse")

WorkingPS<-readRDS("Processed_data/WorkingPS.rds")

set.seed(1024)
oom_diff_res <- diff_analysis(obj = WorkingPS,
                       classgroup = "seq_tech",
                       mlfun = "lda",
                       ratio=0.80,
                       filtermod = "pvalue",
                       firstcomfun = "kruskal.test",
                       padjust = "fdr", #“holm”, “hochberg”, “hommel”, “bonferroni”, “BH”, “BY”, “fdr”, “none”
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       clmin = 10,
                       subclmin = 5,
                       subclwilc = TRUE,
                       ldascore = 4,
                       secondalpha = 0.01,
                       lda = 4,
                       normalization = 1e+06,
                       bootnums=99,
                       ci=0.95,
                       type = "species")


oom_ana_results <- oom_diff_res@result

oom_ana_results[c('order', 'Species')] <- str_split_fixed(oom_ana_results$f, '__', 2)

oom_ana_results_S <- subset(oom_ana_results, order == 's')

head(oom_ana_results_S)

```


```{r,  fig.align="center", echo = TRUE, fig.width = 8, fig.height= 6}

ggplot(oom_ana_results_S,
       aes(x=LDAmean, y=reorder(Species, LDAmean)))+
  geom_segment(aes(yend=Species), xend=0, colour= "grey", 
               linetype="dashed")+
  geom_errorbarh(aes(xmin=oom_ana_results_S$LDAlower, 
                      xmax=oom_ana_results_S$LDAupper,
                      height = .2))+
  geom_point(size=2, shape=21, aes(fill=seq_tech))+
  scale_fill_manual(values=c("#990038", 
                             "#01AED9",
                             "#999999"))+
  theme_bw(base_size = 10)+
  theme(panel.grid.major.y = element_blank())+
  theme(legend.position="top")+
  xlab(bquote(Log[10](LDA)))+
  ylab("Indicator species")+
  theme(axis.text.y = element_text(face = "italic"))

```

```{r,  fig.align="center", echo = TRUE, fig.width = 8, fig.height= 6}
diffclade_p <- ggdiffclade(
                   obj=oom_diff_res,
                   alpha=0.03, 
                   linewd=0.15,
                   skpointsize=0.6, 
                   layout="radial",
                   taxlevel=7, 
                   removeUnkown=FALSE
                 ) +
               scale_fill_manual(
                   values=c("#990038", 
                             "#01AED9",
                             "#999999")
               ) +
               guides(color = guide_legend(
                                  keywidth = 0.1, 
                                  keyheight = 0.6,
                                  order = 3,
                                  ncol=1)
               ) +
               theme(
                   panel.background=element_rect(fill=NA),
                   legend.position="right", 
                   plot.margin=margin(0,0,0,0),
                   legend.spacing.y=unit(0.02, "cm"), 
                   legend.title=element_text(size=7),
                   legend.text=element_text(size=6), 
                   legend.box.spacing=unit(0.02,"cm")
               )
diffclade_p
```
